+++ firefox-ios/Client/Frontend/Browser/BrowserViewController/Actions/GeneralBrowserAction.swift
@@ -87,7 +87,7 @@ enum GeneralBrowserActionType: ActionType {
     case didTapReaderModeBarSummarizerButton
     case shakeMotionEnded
     case showTranslationLanguagePicker
-    case showGoogleLensPhotoPicker
+    case showPhotoPicker
 }
 
 struct GeneralBrowserMiddlewareAction: Action {
