+++ firefox-ios/Client/Frontend/Browser/BrowserViewController/Extensions/BrowserViewController+PHPickerViewControllerDelegate.swift
@@ -1,0 +1,26 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Photos
+import PhotosUI
+import UIKit
+
+extension BrowserViewController {
+    func presentGoogleLensPhotoPicker() {
+        var configuration = PHPickerConfiguration(photoLibrary: .shared())
+        configuration.filter = .images
+        configuration.selectionLimit = 1
+        configuration.preferredAssetRepresentationMode = .current
+
+        let picker = PHPickerViewController(configuration: configuration)
+        picker.delegate = self
+        present(picker, animated: true)
+    }
+}
+
+extension BrowserViewController: PHPickerViewControllerDelegate {
+    func picker(_ picker: PHPickerViewController, didFinishPicking _: [PHPickerResult]) {
+        picker.dismiss(animated: true)
+    }
+}
